server and front connected

+++ account/views.py
@@ -10,7 +10,8 @@ from django.utils.encoding import force_bytes
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from rest_framework.authtoken.models import Token
-# for sending email
+from rest_framework import status
+from django.contrib.auth import logout
 from django.core.mail import EmailMultiAlternatives
 from django.template.loader import render_to_string
 from django.shortcuts import redirect
@@ -38,7 +39,7 @@ class UserRegistrationApiView(APIView):
             print("token ", token)
             uid = urlsafe_base64_encode(force_bytes(user.pk))
             print("uid ", uid)
-            confirm_link = f"http://127.0.0.1:8000/account/active/{uid}/{token}"
+            confirm_link = f"https://lilyloom.onrender.com/account/active/{uid}/{token}"
             email_subject = "Confirm Your Email"
             email_body = render_to_string('confirm_email.html', {'confirm_link' : confirm_link})
             
@@ -91,8 +92,17 @@ def profile(request):
     return Response(serializer.data)
 
 class UserLogoutView(APIView):
+    def post(self, request):
+        if request.user.is_authenticated:
+            # Delete the authentication token if the user is authenticated
+            request.user.auth_token.delete()
+            logout(request)
+            return Response(status=status.HTTP_200_OK)
+        else:
+            # Return a 401 Unauthorized response if the user is not authenticated
+            return Response({"detail": "User is not authenticated."}, status=status.HTTP_401_UNAUTHORIZED)
+
     def get(self, request):
-        request.user.auth_token.delete()
-        logout(request)
+        # Redirect to the login page for GET requests
         return redirect('login')
